Move VM legacy parser regression into root tests
